Setup publishing (#2)
The goal of this is to have access to a published version of the LSP to maven central to facilitate consumption in IDEs.  

Publishing occurs upon merges to the dss branch, or upon tagged release. 

* Adds a mill build (used only for the sake of publishing) 
* Adds dynver-like versioning logic 
* Adds a publish script 
* Adds necessary CI setup 

Publishing is setup correctly, as proven by [this build step](https://github.com/disneystreaming/smithy-language-server/runs/4273043861?check_suite_focus=true). 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.