Expose validator name to use in suppressions
Smithy support suppressing warnings via a metadata. To do
so, add a block like that before the namespace declaration in
your Smithy file:

```
metadata suppressions = [
    {
        id: "UnreferencedShape",
        namespace: "foo.baz",
        reason: "This is a test namespace."
    }
]
```

Right now, diagnostics in the LSP do not report the id of the
validator. It makes it harder for users to suppress a warning
they would like to be gone.

This PR adds the ID in the diagnostic message reported by the
LSP.