More gracefully handle unknown traits
This commit allows hover and jump to definition to continue functioning
when there are unknown traits in the model, and stops crashes when hovering
when there are unknown traits in the model. Additionally, when hovering
over shapes that use an unknown trait, an error message will be displayed in
the hover content. Tests were also added to verify that:
- A model with unknown traits can be loaded
- Diagnostics are still sent for unknown traits
- Definition for valid shapes still works when there are unknown traits
- Hover for valid shapes still works when there are unknown traits
- Error messages are displayed in hover content for invalid shapes