Refactor for performance improvements
Re-writes almost everything in the language server to improve performance,
and lay the ground work for further progress and features. Given the scope
of these changes, this should be considered a WIP as I may have broken
some things.

Overview of performance improvements:
- Per-file model updates on change events
- Model validation only run on save
- Async execution of model building and validation with cancellation
- Async execution of some language features like completion and
document symbol with cancellation
- Incremental file change updates
- Reduced file reads from disk and string copies

From the end user's perspective, only one major change has been made what
the language server can do. It now uses a smithy-build.json in the root
of the workspace as the source of truth for what is part of the project.
Previously, the server loaded all Smithy files it found in any subdir of
the root. This doesn't scale well with multi-root workspaces, and leads to
an issue where Smithy files in the build directory are added to the project,
duplicating sources. The new process looks for a smithy-build.json and uses
only its `sources` and `imports` as files to load into the model (maven deps
are still supported, this is just referring to project files). For backward
compatibility, the old SmithyBuildExtensions `build/smithy-dependencies.json`
and `.smithy.json` are still supported. A new file, `.smithy-project.json`,
is being developed which allows projects that are configured outside of a
smithy-build.json (such as a Gradle project) to specify their project files
_and_ dependencies. Right now these dependencies are local, paths to JARs,
but it may make sense to support Maven dependencies in there as well. More
to come on how `.smithy-project.json` works in documentation updates. To
support using the language server without a smithy-build.json, a future
update is in progress to allow a 'detached' project which loads whatever
files you open.

Other updates:
- Use smithy-syntax formatter
- Report progress to client on load
- Add configuration option for the minimum severity of validation events
- Update dependencies