remove metadata when reloading single files
The performance refactor caused a regression where making changes to a file
with metadata would cause that metadata key to be duplicated. Since the
server used to reload all files on a change, we never had to worry about this,
but since we're trying to now only reload single files, we need to remove
any metadata in that file from the model before reloading.

This works similarly to how we collect per-file shapes, but is slightly more
complex because array metadata with the same key get merged into a single
array (as opposed to non-arrays, which cause an error when there are the same
keys).