Properly load detached files with invalid syntax
Fixes an issue where basically any time you created a file not
attached to a project, it wouldn't ever be loaded properly and
any updates wouldn't register. This happened because the server
wasn't creating a SmithyFile for a detached file if it had no
shapes in it. The SmithyFile is created only when the project is
initialized, so subsequent changes wouldn't fix it.