Fix applying traits in partial load
The partial loading strategy, which is meant to reduce the amount
of unnecessary re-parsing of unchanged files on every file change,
works by removing shapes defined in the file that has changed. But
this wasn't taking into account traits applied using `apply` in
other files. When a shape is removed, all the traits are removed,
and the `apply` isn't persisted. So we need to keep track of all
trait applications that aren't in the same file as the shape def.
Additionally, list traits have their values merged, so we actually
need to either partially remove a trait (i.e. only certain elements), or
just reload all the files that contain part of the trait's value. This
implementation does the latter, which also requires creating a sort
of dependency graph of which files need other files to be loaded with
them. There's likely room for optimization here (potentially switching
to the first approach), but we will have to guage the performance.

This commit also consolidates the project updating logic for adding,
removing, and changing files into a single method of Project, and
adds a bunch of tests around different situations of `apply`.