Fix Document, UriAdapter, and tests for windows
Previously, Document used System.lineSeparator() for figuring out
where line starts would be (index of linesep + 1). But if the file
was created (and, say, packaged in a jar) on another OS, it would
have different lineseps. This change makes use of a simple fact I
overlooked in the initial implementation, which was that '\n' is
still the last character on each line, so we don't need to break
on System.lineSeparator(), just on newline (unless there's still
some OS using '\r' only line breaks).

UriAdapter was updated to handle windows URIs, which would be made
into invalid paths with a leading '/' after removing 'file://'.

A bunch of test cases were also updated, which essentially all had
one or both of the above problems.