Upgrade dependencies and jdk to 21 (#157)
* Upgrade dependencies and jdk to 21

A few minor code changes were made for the upgrade, including removing
LspLog (which should have already been removed). I didn't make any larger
changes to make use of java 21 features - that can be done later. The
primary motivation of this commit is to update our dependency on lsp4j,
which outdated and depended on a version of guava with two CVEs out:
CVE-2023-2976, CVE-2020-8908. In order to do so, a jdk upgrade was
required since newer versions of lsp4j are past jdk 8. We were going
to upgrade the jdk soon anways.

Also got rid of an extra signing block in the publishing task, which
is taken care of by jreleaser.