Fix rebuilding when SourceLocation is NONE (#196)
The rebuild index checks which shapes have traits that are applied in
other files (i.e. via an `apply`), so we can preserve those traits in a
reload. However, if a trait's source location is `SourceLocation.NONE`,
which can happen if it wasn't set on the trait's builder, the rebuild
index would consider that "a trait applied from another file". If the
trait isn't actually being applied from another file, this would cause
a duplicate trait conflict, since the trait would be added both from the
rebuild index, and the ModelAssembler reparsing.

This commit changes two things:
1. The trait's _node_ source location is now used for the comparison,
   since it should always have a source location as it is constructed
   within smithy-model, not by trait provider implementations
2. This source location is checked to see if it is NONE.