Cleanup some noisy logs and debugging in tests
Logging for unknown file in ServerState is unnecessary because it will
always log for newly opened detached files, plus callers have to handle
a file not being found anyway.

Logging for deprecated properties in SmithyBuildExtensions aren't going
to be visible enough to customers to be useful, plus we emit a warning
for build exts anyways.

I also had some old debugging code lingering in tests that I cleaned up.