Fix parse errors for empty build files
Document::rangeBetween would return `null` if the document was empty,
but I wasn't checking for that in ToSmithyNode (which creates parse
events).

The reason the range is `null` is because Document.lineOfIndex returns
oob for an index of 0 into an empty document. Makes sense, as an empty
document has no lines. I updated a DocumentTest to clarify this
behavior.