Add hover for non-shape keywords, and more links
For non-shape keywords, I added a new shape to the builtins model,
NonShapeKeywords. I also had to update Statements so you can check if
the cursor is within the keyword because some statements don't need to
store the keyword Ident. Then I updated IdlPosition to perform this
check.

I also added way to include some externalDocumentation links to the
hover content of all members in a builtin shape, with the intent that
members which don't have a doc link themselves will still have some link
people can use to get to the Smithy site if they need more detail that
isn't provided in the hover content. The way this works is, if the
builtin shape you're getting the hover content for is a member, we also
append the externalDocumentation defined on its container shape. This
way we don't have to repeat the trait on every member.