Add missing hover docs (#204)
This adds hover info for a bunch of things that I didn't get around to in the initial hover upgrade,
including:

All shape types
Non shape type keywords, metadata, namespace, use, apply, with, and for
Member names in service, resource, and operation shapes
Links to Smithy docs site for all these "builtins"
Documentation traits on user-defined members (before it just wasn't rendered)
I had to add new builtin model for the non-shape keywords, but reused ShapeMemberTargets for
shape type keyword hover.

I also had to add a way to check if the cursor was inside the keyword of a statement, which simplifies
IdlPosition a bit too.

The Smithy docs links use externalDocumentation, like the smithy-build.json model, and I also added
a way to include docs links to all members of a builtin shape, because there's some things which don't
have a great link for themselves specifically, but it would still be nice to allow navigating to some
docs page that could be useful. This works by adding the externalDocumentation trait to the root
shape in the builtins model, and having the hover implementation add any external docs from the
container shape when constructing hover for a builtin. So, for example, hovering over "version": "1.0"
in smithy-build.json now also includes a link to the smithy-build.json page.